pyproject.toml
[tool.poetry]
name = "captain"
version = "0.1.0"
description = ""
authors = ["noah <noah@apartalk.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "3.25.0"
psycopg2-binary = "*"
importlib-metadata = "*"
inject = "*"
pydantic = "*"
pypubsub = "*"
flasgger = "*"
Authlib = "^0.15.3"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest-xdist = "*"
pytest = "*"
pytest-factoryboy = "*"
faker = "*"
python-dotenv = "^0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
